cashManagementUi/src/test/javascript/spec/app/entities/cm-error/cm-error.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { CmErrorService } from 'app/entities/cm-error/cm-error.service';
import { ICmError, CmError } from 'app/shared/model/cm-error.model';

describe('Service Tests', () => {
    describe('CmError Service', () => {
        let injector: TestBed;
        let service: CmErrorService;
        let httpMock: HttpTestingController;
        let elemDefault: ICmError;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(CmErrorService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new CmError(0, 0, 0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', currentDate);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        errornDateTime: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a CmError', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        errornDateTime: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        errornDateTime: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new CmError(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a CmError', async () => {
                const returnedFromService = Object.assign(
                    {
                        errorId: 1,
                        errorComponent: 1,
                        errorCode: 'BBBBBB',
                        errorDescription: 'BBBBBB',
                        errorStackTrace: 'BBBBBB',
                        errornDateTime: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        errornDateTime: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of CmError', async () => {
                const returnedFromService = Object.assign(
                    {
                        errorId: 1,
                        errorComponent: 1,
                        errorCode: 'BBBBBB',
                        errorDescription: 'BBBBBB',
                        errorStackTrace: 'BBBBBB',
                        errornDateTime: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        errornDateTime: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a CmError', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
